% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basis_catch.R
\name{get_basis_catch}
\alias{get_basis_catch}
\title{get_basis_catch}
\usage{
get_basis_catch(
  start_year = 2021,
  end_year = 2021,
  min_lat = 50,
  max_lat = 80,
  tsn
)
}
\arguments{
\item{tsn}{taxonomic serial number}

\item{startyear}{first year in the time series, default 2021}

\item{endyear}{last year in the time series, default 2021}

\item{minlat}{minimum latitude, used to differentiate NBS/SEBS, default 50}

\item{maxlat}{maximum latitude, used to differentiate NBS/SEBS, default 80}
}
\description{
This function pulls basis survey catch data
by species and life history stage including zero catches.
All parameters are optional but downloading all years, all species may take several minutes.
The default is to download only 2021.
A related function, get_tsn(), pulls a list of tsns to lookup.
}
\examples{
##get catch from 2004
data<-get_basis_catch(startyear=2004, endyear=2004)

##get catch from 2004 from the NBS
data<-get_basis_catch(startyear=2004, endyear=2004, minlat=60)

##Retrieve Pacific cod data from 2004 South of 60 degrees North
#look up tsn from get_tsn function
get_tsn()\%>\%
filter(commonname=="Pacific Cod")
#pull data
data<-get_basis_catch(startyear=2004, endyear=2004, maxlat=60, tsn=164711)
}
